# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiError": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "details": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RevisionConflictDetails"
              },
              {
                "type": "null"
              }
            ]
          },
          "field_errors": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "title": "Field Errors",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "message_args": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Message Args",
            "type": "object"
          },
          "retryable": {
            "default": false,
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ApiError",
        "type": "object"
      },
...
      "CollectionCreate": {
        "additionalProperties": false,
        "properties": {
          "card_id": {
            "format": "uuid",
            "title": "Card Id",
            "type": "string"
          },
          "card_revision": {
            "minimum": 1.0,
            "title": "Card Revision",
            "type": "integer"
          },
          "confirmed_target_language": {
            "anyOf": [
              {
                "enum": [
                  "ja",
                  "en"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmed Target Language"
          },
          "notebook_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 20,
            "title": "Notebook Ids",
            "type": "array"
          }
        },
        "required": [
          "card_id",
          "card_revision"
        ],
        "title": "CollectionCreate",
        "type": "object"
      },
      "CollectionRead": {
        "additionalProperties": false,
        "properties": {
          "card_id": {
            "format": "uuid",
            "title": "Card Id",
            "type": "string"
          },
          "card_revision": {
            "minimum": 1.0,
            "title": "Card Revision",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_text": {
            "title": "Display Text",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "const": "word",
            "default": "word",
            "title": "Kind",
            "type": "string"
          },
          "payload": {
            "$ref": "#/components/schemas/WordCardPayload"
          },
          "revision": {
            "minimum": 1.0,
            "title": "Revision",
            "type": "integer"
          },
          "source_refs": {
            "items": {
              "$ref": "#/components/schemas/NovelContentLocator"
            },
            "title": "Source Refs",
            "type": "array"
          },
          "target_language": {
            "enum": [
              "ja",
              "en"
            ],
            "title": "Target Language",
            "type": "string"
          }
        },
        "required": [
          "id",
          "revision",
          "card_id",
          "card_revision",
          "display_text",
          "target_language",
          "payload",
          "source_refs",
          "created_at"
        ],
        "title": "CollectionRead",
        "type": "object"
      },
...
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiError"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "error",
          "meta"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "ExplanationResolveRead": {
        "additionalProperties": false,
        "properties": {
          "results": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ResolvedCard"
                },
                {
                  "$ref": "#/components/schemas/MissingCard"
                }
              ]
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "results"
        ],
        "title": "ExplanationResolveRead",
        "type": "object"
      },
      "ExplanationResolveRequest": {
        "additionalProperties": false,
        "properties": {
          "targets": {
            "items": {
              "$ref": "#/components/schemas/ResolveTarget"
            },
            "maxItems": 10,
            "minItems": 1,
            "title": "Targets",
            "type": "array"
          }
        },
        "required": [
          "targets"
        ],
        "title": "ExplanationResolveRequest",
        "type": "object"
      },
...
      "PageMeta": {
        "additionalProperties": false,
        "properties": {
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "next_cursor",
          "has_more"
        ],
        "title": "PageMeta",
        "type": "object"
      },
      "PageResponse_CollectionRead_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/CollectionRead"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/PageMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "PageResponse[CollectionRead]",
        "type": "object"
      },
      "PageResponse_MaterialSummary_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/MaterialSummary"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/PageMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "PageResponse[MaterialSummary]",
        "type": "object"
      },
...
      "ResolveTarget": {
        "additionalProperties": false,
        "properties": {
          "explanation_language": {
            "maxLength": 35,
            "minLength": 2,
            "title": "Explanation Language",
            "type": "string"
          },
          "source_locator": {
            "$ref": "#/components/schemas/NovelContentLocator"
          },
          "target_language": {
            "enum": [
              "ja",
              "en"
            ],
            "title": "Target Language",
            "type": "string"
          }
        },
        "required": [
          "source_locator",
          "target_language",
          "explanation_language"
        ],
        "title": "ResolveTarget",
        "type": "object"
      },
...
      "ResponseMeta": {
        "additionalProperties": false,
        "properties": {
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "request_id"
        ],
        "title": "ResponseMeta",
        "type": "object"
      },
...
      "SuccessResponse_CollectionRead_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CollectionRead"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "SuccessResponse[CollectionRead]",
        "type": "object"
      },
...
      "SuccessResponse_ExplanationResolveRead_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ExplanationResolveRead"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "SuccessResponse[ExplanationResolveRead]",
        "type": "object"
      },
...
      "SuccessResponse_NovelChapterRead_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/NovelChapterRead"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "SuccessResponse[NovelChapterRead]",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Haruka API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/collections": {
      "get": {
        "operationId": "list_collections",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponse_CollectionRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Collections",
        "tags": [
          "learning-reference"
        ],
        "x-haruka-permissions": [
          "client.collection.read"
        ]
      },
      "post": {
        "operationId": "create_collection",
        "parameters": [
          {
            "in": "header",
            "name": "idempotency-key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollectionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_CollectionRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Collection Create",
        "tags": [
          "learning-reference"
        ],
        "x-haruka-permissions": [
          "client.collection.create",
          "client.material.read"
        ]
      }
    },
    "/api/v1/explanations/resolve": {
      "post": {
        "operationId": "resolve_explanations",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExplanationResolveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_ExplanationResolveRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Explanation Resolve",
        "tags": [
          "learning-reference"
        ],
        "x-haruka-permissions": [
          "client.ai.explain",
          "client.material.read"
        ]
      }
    },
...
    "/api/v1/materials": {
      "get": {
        "operationId": "list_materials",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponse_MaterialSummary_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Materials",
        "tags": [
          "learning-reference"
        ],
        "x-haruka-permissions": [
          "client.material.list"
        ]
      }
    },
...
    "/api/v1/novels/{material_id}/revisions/{revision_id}/chapters/{node_id}": {
      "get": {
        "operationId": "get_novel_chapter",
        "parameters": [
          {
            "in": "path",
            "name": "material_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Material Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Revision Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Node Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_NovelChapterRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Novel Chapter",
        "tags": [
          "learning-reference"
        ],
        "x-haruka-permissions": [
          "client.material.read"
        ]
      }
    },
...
  }
}